Improves test message prefix.
It now links to the build in the current repo, even if that's not the default.

It also uses ChatExchange@localhost, linked to the git repo, when not
being run on Travis. (I wasn't sure what label to use here.)

Previous, it would generate an invalid or weird Travis link in these cases.