Adds examples/web_viewer.py, a simple browser-based chat viewer.
We're not building a JavaScript application here, so we are never going
to put any state exclusively in the JavaScript; it's just going to be
some Angular templates reflecting state that's frequently updated from
the server.

We could add inputs, later, but those should just immediately post to
our server, they shouldn't touch the client-side state.