Fix the BS4 search for the script needed to count current user IDs
Stack Exchange has apparently added another <script> element to the
header of the chat page, so when Browser.get_current_user_ids_in_room()
searches for the fourth script element which contains the list of
users present, it find a script in the head, not the one it's supposed
to find in the body of the page. This commit fixes that by having it
search for the first <script> in the body, which should be more resilient
against future changes in the scripts used in the header. The same
fix is applied to Browser.get_current_user_names_in_room().