Fix bug when logging in
The user ID is retrieved by parsing the user URL.  When using / as the
delimiter, the user ID is at position 2 if the URL is of the form
stackoverflow.com/user/0000/username, but will be at position 4 if the
URL is of the form https://stackoverflow.com/user/0000/username.
To fix this problem, find the ID from the end rather than from the
beginning.  My previous patch didn't account for the username at the
end of the URL, so please ignore that patch.